examples/latcoder/requirements.txt
playwright
Pillow
openai==1.55.3
pyarrow
datasets
google-generativeai==0.8.3
lap
nltk
scikit-image
rouge
opencv-python
beautifulsoup4
scikit-learn
clip@git+https://github.com/openai/CLIP.git@main
graphviz
SwissArmyTransformer==0.4.9
deepseek_vl@git+https://github.com/deepseek-ai/DeepSeek-VL.git@main
https://vllm-wheels.s3.us-west-2.amazonaws.com/nightly/vllm-1.0.0.dev-cp38-abi3-manylinux1_x86_64.whl
transformers==4.45.2
easyocr
open_clip_torch
google-genai
